+++ src/components/layout/component.jsx
@@ -11,7 +11,7 @@ export const Layout = ({ children }) => {
                 <Footer />
         
             <div id="modal-container" />
-            <div id="popover-container"className={styles.popoverContainer}/>
+            <div id="popover-container" className={styles.popoverContainer}/>
         </div>
     )
 }
