+++ src/components/modal-loading/component.jsx
@@ -1,0 +1,13 @@
+import { createPortal } from "react-dom";
+import styles from "./styles.module.scss";
+
+export const ModalIsLoading = () => {
+    const modalContainer = document.getElementById("modal-container");
+    return createPortal(
+        <div className={styles.root}>
+            is loading...
+        </div>,
+        modalContainer
+    );
+};
+
